third_party/WebKit/LayoutTests/media/media-play-promise.html

<!DOCTYPE html>
<title>Test the play() behaviour with regards to the returned promise for media elements.</title>
<script src="../resources/testharness.js"></script>
<script src="../resources/testharnessreport.js"></script>
<script src="media-file.js"></script>
<script>
// This is testing the behavior of play() with regards to the returned
// promise. This test file is creating a small framework in order to be able
// to test different cases easily and independently of each other.
// All tests have to be part of the tests and testsWithUserGesture arrays.
// Each test returns a promise for audio.play() which is either resolved or rejected.

var tests = [
    // Test that play() on an element that doesn't have enough data will
    // return a promise that resolves successfully.
    function playBeforeCanPlay(t, audio) {
        audio.src = findMediaFile("audio", "content/test");
        assert_equals(audio.readyState, HTMLMediaElement.HAVE_NOTHING);
        playExpectingResolvedPromise(t, audio);
    },

    // Test that play() on an element that has enough data will return a
    // promise that resolves successfully.
    function playWhenCanPlay(t, audio) {
        audio.src = findMediaFile("audio", "content/test");

        audio.oncanplay = t.step_func(function() {
            assert_greater_than_equal(audio.readyState, HTMLMediaElement.HAVE_FUTURE_DATA);
            assert_true(audio.paused);
            playExpectingResolvedPromise(t, audio);
        });
    },

    // Test that play() on an element that is already playing returns a
    // promise that resolves successfully.
    function playAfterPlaybackStarted(t, audio) {
        audio.preload = "auto";
        audio.src = findMediaFile("audio", "content/test");

        audio.onplaying = t.step_func(function() {
            assert_equals(audio.readyState, HTMLMediaElement.HAVE_ENOUGH_DATA);
            assert_false(audio.paused);
            playExpectingResolvedPromise(t, audio);
        });

        audio.oncanplaythrough = t.step_func(function() {
            audio.play();
        });
    },

    // Test that play() on an element with an unsupported content will
    // return a rejected promise.
    function playNotSupportedContent(t, audio) {
        audio.src = findMediaFile("audio", "data:,");

        audio.onerror = t.step_func(function() {
            assert_true(audio.error instanceof MediaError);
            assert_equals(audio.error.code, MediaError.MEDIA_ERR_SRC_NOT_SUPPORTED);
        });
        playExpectingRejectedPromise(t, audio, "NotSupportedError");
    },

    // Test that play() returns a resolved promise if called after the
    // element suffered from a decode error.
    // This test doesn't test a spec behaviour but tests that the Blink
    // implementation properly changed after the spec changed.
    function playDecodeError(t, audio) {
        audio.src = findMediaFile("audio", "content/test");

        audio.onerror = t.step_func(function() {
            assert_true(audio.error instanceof MediaError);
            assert_equals(audio.error.code, MediaError.MEDIA_ERR_DECODE);
        });

        // The setMediaElementNetworkState() method requires metadata to be available.
        audio.onloadedmetadata = t.step_func(function() {
          internals.setMediaElementNetworkState(audio, 6); // NetworkStateDecodeError.
          playExpectingResolvedPromise(t, audio);
        });
    },

    // Test that play() returns a resolved promise if called after the
    // element suffered from a network error.
    // This test doesn't test a spec behaviour but tests that the Blink
    // implementation properly changed after the spec changed
    function playNetworkError(t, audio) {
        audio.src = findMediaFile("audio", "content/test");

        audio.onerror = t.step_func(function() {
            assert_true(audio.error instanceof MediaError);
            assert_equals(audio.error.code, MediaError.MEDIA_ERR_NETWORK);
        });

        // The setMediaElementNetworkState() method requires metadata to be available.
        audio.onloadedmetadata = t.step_func(function() {
          internals.setMediaElementNetworkState(audio, 5); // NetworkStateNetworkError.
          playExpectingResolvedPromise(t, audio);
        });
    },

    // Test that play() returns a rejected promise if the element is
    // sufferring from a not supported error.
    function playWithErrorAlreadySet(t, audio) {
        audio.src = findMediaFile("audio", "data:,");

        audio.onerror = t.step_func(function() {
            assert_true(audio.error instanceof MediaError);
            assert_equals(audio.error.code, MediaError.MEDIA_ERR_SRC_NOT_SUPPORTED);
            playExpectingRejectedPromise(t, audio, "NotSupportedError");
        });
    },

    // Test that play() returns a resolved promise if the element had its
    // source changed after suffering from an error.
    function playSrcChangedAfterError(t, audio) {
        audio.src = findMediaFile("audio", "data:,");

        audio.onerror = t.step_func(function() {
            assert_true(audio.error instanceof MediaError);
            assert_equals(audio.error.code, MediaError.MEDIA_ERR_SRC_NOT_SUPPORTED);
            audio.src = findMediaFile("audio", "content/test");

            audio.onloadedmetadata = t.step_func(function() {
                playExpectingResolvedPromise(t, audio);
            });
        });
    },

    // Test that play() returns a rejected promise if the element had an
    // error and just changed its source.
    function playRaceWithSrcChangeError(t, audio) {
        audio.src = findMediaFile("audio", "data:,");

        audio.onerror = t.step_func(function() {
            assert_true(audio.error instanceof MediaError);
            assert_equals(audio.error.code, MediaError.MEDIA_ERR_SRC_NOT_SUPPORTED);
            audio.src = findMediaFile("audio", "content/test");
            assert_equals(audio.error, null);
            assert_equals(audio.readyState, HTMLMediaElement.HAVE_NOTHING);
            playExpectingResolvedPromise(t, audio);
        });
    },

    // Test that play() returns a resolved promise when calling play() then
    // pause() on an element that already has enough data to play. In other
    // words, pause() doesn't cancel play() because it was resolved
    // immediately.
    function playAndPauseWhenCanplay(t, audio) {
        audio.src = findMediaFile("audio", "content/test");

        audio.oncanplaythrough = t.step_func(function() {
            assert_equals(audio.readyState, HTMLMediaElement.HAVE_ENOUGH_DATA);
            playExpectingResolvedPromise(t, audio);
            assert_false(audio.paused);
            audio.pause();
            assert_true(audio.paused);
        });
    },

    // Test that play() returns a rejected promise when calling play() then
    // pause() on an element that doesn't have enough data to play. In other
    // words, pause() cancels play() before it can be resolved.
    function playAndPauseBeforeCanPlay(t, audio) {
        assert_equals(audio.readyState, HTMLMediaElement.HAVE_NOTHING);
        playExpectingRejectedPromise(t, audio, "AbortError");
        assert_false(audio.paused);
        audio.pause();
        assert_true(audio.paused);
    },

    // Test that load() rejects all the pending play() promises.
    // This might be tested by other tests in this file but it is present in
    // order to be explicit.
    function loadRejectsPendingPromises(t, audio) {
        playExpectingRejectedPromise(t, audio, "AbortError"); // the promise will be left pending.
        audio.load();
    },

    // Test that changing the src rejects the pending play() promises.
    function newSrcRejectPendingPromises(t, audio) {
        playExpectingRejectedPromise(t, audio, "AbortError"); // the promise will be left pending.
        audio.src = findMediaFile("audio", "content/test");
    },

    // Test ordering of events and promises.
    // This is testing a bug in Blink, see https://crbug.com/587871
    function testEventAndPromiseOrdering(t, audio) {
        audio.src = findMediaFile("audio", "data:,");

        audio.onerror = t.step_func(function() {
            // Until https://crbug.com/587871 is fixed, the events will be
            // [ promise, volumechange, volumechange ], it should be
            // [ volumechange, promise, volumechange ]. Right now test finishes
            // as soon as promise is rejected, before "volumechange" is fired.
            var numVolumeChangeEvents = 0;
            audio.onvolumechange = t.step_func(function() { ++numVolumeChangeEvents; });
            audio.volume = 0.1;
            audio.play().then(t.unreached_func(), t.step_func_done(function() {
                assert_equals(arguments.length, 1);
                assert_equals(arguments[0].name, "NotSupportedError");
                assert_equals(numVolumeChangeEvents, 0);
            }));
            audio.volume = 0.2;
        });
    },

    // Test that calling pause() then play() on an element that is already
    // playing returns a promise that resolves successfully.
    function pausePlayAfterPlaybackStarted(t, audio) {
        audio.preload = "auto";
        audio.src = findMediaFile("audio", "content/test");

        audio.onplaying = t.step_func(function() {
            assert_equals(audio.readyState, HTMLMediaElement.HAVE_ENOUGH_DATA);
            assert_false(audio.paused);
            audio.pause();
            playExpectingResolvedPromise(t, audio);
        });

        audio.oncanplaythrough = t.step_func(function() {
            audio.play();
        });
    }
];

tests.forEach(function(test) {
    internals.settings.setMediaPlaybackRequiresUserGesture(false);
    async_test(function(t) {
        var audio = document.createElement("audio");
        test(t, audio);
    }, test.name);
});

var testsWithUserGesture = [
    // Test that play() on an element when media playback requires a gesture
    // returns a resolved promise if there is a user gesture.
    function playRequiresUserGestureAndHasIt(t, audio) {
        audio.src = findMediaFile("audio", "content/test");
        playWithUserGesture(t, audio);
    },

    // Test that play() on an element when media playback requires a gesture
    // returns a rejected promise if there is no user gesture.
    function playRequiresUserGestureAndDoesNotHaveIt(t, audio) {
        audio.src = findMediaFile("audio", "content/test");
        playExpectingRejectedPromise(t, audio, "NotAllowedError");
    }
];

testsWithUserGesture.forEach(function(test) {
    internals.settings.setMediaPlaybackRequiresUserGesture(true);
    async_test(function(t) {
        var audio = document.createElement("audio");
        test(t, audio);
    }, test.name);
});

function playExpectingResolvedPromise(t, audio) {
    audio.play().then(t.step_func_done(function() {
        assert_equals(arguments.length, 1);
        assert_equals(arguments[0], undefined);
    }), t.unreached_func());
}

function playExpectingRejectedPromise(t, audio, error) {
    audio.play().then(t.unreached_func(), t.step_func_done(function() {
        assert_equals(arguments.length, 1);
        assert_equals(arguments[0].name, error);
    }));
}

function playWithUserGesture(t, audio) {
    document.onclick = function() {
        playExpectingResolvedPromise(t, audio);
        document.onclick = null;
    };

    eventSender.mouseDown();
    eventSender.mouseUp();
}
</script>
